+++ packages/terra-responsive-element/tests/jest/ResponsiveElement.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ResponsiveElement from '../../src/ResponsiveElement';
 
-it('should render a default component', () => {
-  const responsiveElement = shallow(<ResponsiveElement defaultElement={<div>Default</div>} />);
+it('should render the component defined for tiny', () => {
+  const responsiveElement = shallow(<ResponsiveElement tiny={<div>Tiny</div>} />);
   expect(responsiveElement).toMatchSnapshot();
 });